requirements.txt
discord.py[voice]
pyyaml